print("select")+addon prefs
I added a print("selection") -> Ctrl+Alt+P (this is just adding a bool to the simple print class to call the operator with the prop enabled or not)
In addon_prefs I added a label with shortcuts, and a bool to enable or not get click areas (default= True). 
not hurry at all, but I push it if not I won't remember later I did it.