new copy method
- if selection, copy the selection
- if no selection copy word under cursor
- if no selection and no word under cursor copy what is in the clipboard
  so this is possible to copy something go to an empty text zone and do e.g print

this method of copy allows to do more things on text and selection. but the way blender editor is working in text editor is weird selection+ cursor... 
under console I did this https://github.com/1C0D/console_easy_text_edit-Blender this is linking the cursor automatically to the selection so the console is behaving exactly like a text editor. and there is undo redo :)